Initial Commit of the Beginning of File Hashing Application.
This project is already of considerably larger scope than my last weekend projects! The most interesting part (and also the part I dived into first) was the hashing object that would be running as a separate thread. I knew I'd need a thread pool and an object to represent a hashing job for each thread running, to maximize how quickly this app could hash multiple files. The communication between the running task and the GUI thread needed some work but Qt makes this pretty easy with Queued signals and slots.

Unlike the last two projects I am trying to embrace a more test driven approach and have set up testing along with this. The testing is barebones, but so is the program itself, and other than the small, unused GUI stuff everything currently does and will have test coverage. The HashTasksModel doesn't really have tests, but it doesn't really exist either. The HashingJob class needs its own testing, but that class also doesn't exist yet so there's nothing to test in the first place!

It was just time to make a commit.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>