Second Major Commit of this Project.
It's mostly about making the test bed and getting the Hashing functionality fleshed out. The testing is leading this project so I am making sure that everything underneath is operating properly before going further. There's a lot more to do and this project is not really close to done, but the backend is much better. The Model object that works with the HashingJob objects is next, along with its test bed, and then the GUI and everything that goes with that.

The communication seems pretty solid and the hashing is good. Dependency on Crypto++ seems to be working out well. I'd like to add more hash algorithms later but this is plenty until the GUI and everything is sorted out.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>