Complete Revision of the HashTask system, and got basic functionality in the HashTasksModel object.
Tests are mostly done, and a small start has been made with the Item Delegate class that will be displaying the loading bar for all of the HashTask objects. Things have been delayed for a while but I've gotten back into this. Project is going along well and after I finish the HashTasksModel and its according testing I will merge this branch.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>