The model basically works, so it is time to merge
The HashTasksModel is now in a working state and it now updates the screen automatically for how far it is going. Things appear to be working, however I have found a critical bug where it does not hash very large files. That said, I have to get this in place, so here's a merge of a broken version. The bug will be fixed as soon as I can get it done.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>