Stability Commit, Revised HashTask heavily to use PImpl idiom, fleshed out the core features, and made the GUI a little less hideous.
Things are moving along, but I need to make a branch for all the GUI changes I am planning. Now it is time to make the big push into quality-of-life improvements in the software, like the Preferences window, matching against known hashes, and copying the hashes to the clipboard. I should also add more algorithms as soon as I can!

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>