Committing Major Revision of the SimpleFileHash Application.
It has been a while since I made a commit, so this seems like an appropriate enough place to do so. A lot has been added, namely the new FileHashApplication object that is for the application itself as opposed to just putting that in the main() function and creating a MainWindow there. I've implemented the custom MD5 wrapper and made it so it works in this application, and it actually appears to be going a bit faster than the MD5 implementation in Crypto++. The PreferencesDialog has been added, and a UserSettings object to store the settings a user creates.

There are still a lot of things that need implementing. The HashMatchWindow is basically nothing right now, but it should in the long term be a window that allows you to input expected hashes and then highlights them in green when they are found and red when they are not. I've also changed the HashTasksModel to provide better information from the data() method.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>