A ton of updates and changes, added MD6 functionality, added translation files for Spanish, French, Korean, Polish, and Swedish, and updated tests.
So I keep doing a lot of different things to this software to get it off the ground. This is the chaotic moments of creation where things move fast and big changes are commonplace. I've significantly improved the UI experience, got Dark Mode working natively on Windows, and made the PreferencesDialog make sense. The HashMatchWindow is more fleshed out but still isn't functional, so that still needs to be done.

I've also added an sfhbase.hpp file, and put everything under the KirHut::SFH namespace. Namespaces are pretty much always good and never bad, so they are used for this application. The hashalgo.cpp because the Algo type was needing functions that executed at runtime compiled somewhere and I've included a lot of expanded abilities that should be very efficient courtesy of constexpr.

The HashTask object has been updated to include MD6 functionality at 5 bit lengths, 128, 224, 256, 384, and 512. There are no tests currently for 224 or 384 because I haven't found an online 3rd party source to verify the hashes with, so I am leaving those until I do find somewhere that provide me with verified hashes of my random data files.

The HashingJob has been updated with additional functionality, namely when it opens a single file, it will use all available hashing algorithms in the job instead of the user selected algorithm. This makes it so users can quickly hash a single file in a lot of different hashing algorithms.

Massive updates to the UserSettings object, it should now save Geometry QByteArrays that Qt creates from the windows instead of raw values. This allows it to save things like Maximized and other state data that isn't in a basic size value. It should be much easier to save window state data like this as well.

While the translation files have been added and the update_translations target has been run in CMake, the translating work itself has not been done and no translations are actually available yet. This should change by the next update.

There are also some bugs, like rarely hashes will show "Calculating Hash..." instead of the completed hash after it calls finished().

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>