Major UI Update, setup Context Menu Install system, updated all translations, created the meat of the PreferencesDialog functionality, implemented dark and light mode overrides, implemented immediate translation in PreferencesDialog, and fixed tons of bugs.
The biggest improvement here is the PreferencesDialog was mostly implemented. This led the development efforts, including fleshing out the functionality. There is still more that needs to be done on PreferencesDialog, such as implementing the maximum number of files in a HashingJob. HashingJob was updated to provide a new functionality when a single file is found and the passed Algo is Algo::None; now you get all of the implemented functions that are still enabled in the PreferencesDialog.

Drag and Drop was implemented, and it works with the navigateSubdirectories setting. The FileHashApplication object has been considerably updated with new functionality, and is where the Settings object and translations now live. Tests have been updated to reflect new functionality. Implementing proper dark mode and light mode took some time, and will probably not work at all on Mac, but that is for later and should look very good on both Windows and Linux.

All of the translations were added for the main five target languages. This took a lot of busy work but once you are in a groove it is pretty easy. All of the translation work was done with ChatGPT.

There is still a bunch of stuff that is not working or is completely not implemented:
* The settings loading is still a bit janky, there are parts that don't load or save to file properly.
* The entire Hash Matching Window functionality is non-existent. This is the top priority for the next branch.
* The Context Menu Installation system works on absolutely no operating systems at all. I need to implement this at a minimum for Windows, Linux, and Mac.
* Pausing functionality and individual canceling of Hash Jobs should exist.
* There should be a way for the user to set the settings back to default.

There's also probably a whole bunch more bugs to deal with.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>