More UI improvements, finished most HashMatchWindow functionality, added UnmatchedHashModel and HashMatchModel, implemented opening HashDeep files, implemented Uppercase options, fixed a bunch of bugs, and caught up translations.
There is of course a lot that was done here. The AboutDialog was added, along with all of the functionality for the HashMatchWindow. This includes a table for hash matches as well as a table for unmatched hashes, to keep them organized for the user. Opening Hash Sum files and HashDeep files should now work, however this functionality isn't tested and the feature needs to be better integrated.

All of the translations have been updated with new text in the application. There are a few WhatsThis strings that aren't translated yet, but for the most part it is complete. I will need to add more languages to this application in the future, but it should be relatively easy to do for a translator.

The PreferencesDialog has been updated to allow the user to display completed hashes in uppercase instead of lowercase. This is mostly aesthetic, as the underlying string is still lowercase, however when file output is implemented this should matter to the user.

Things that still need to be done:
* I still need to implement the Right-click context menu installation system for all OSs.
* Output of HashDeep and hash sum files still needs to be implemented.
* There should be a fully functional command line interface.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>